intégration d'une barre de recherche en plus de la map
