+++ package.json
@@ -12,7 +12,9 @@
     "react": "16.8.3",
     "react-dom": "^16.8.6",
     "react-native": "https://github.com/expo/react-native/archive/sdk-34.0.0.tar.gz",
-    "react-native-maps": "^0.25.0",
+    "react-native-elements": "^1.1.0",
+    "react-native-maps": "^0.24.2",
+    "react-native-responsive-screen": "^1.2.2",
     "react-native-web": "^0.11.4"
   },
   "devDependencies": {
